Refactor: 캘린더 중복 서비스 로직 분리 #49
