+++ apps/front-end/src/app/(routes)/date-calendar/page.tsx
@@ -1,7 +1,8 @@
-import { redirect } from 'next/navigation';
 import LoginPrompt from '@/components/features/date-calendar/login-prompt';
+import { getMyCouple } from '@/lib/services/temp-couples-server.services';
 import { getServerCookie, getUserIdFromToken } from '@/lib/utils/cookes.utils';
 import { Metadata } from 'next';
+import { redirect } from 'next/navigation';
 
 export const metadata: Metadata = {
   title: '커플 캘린더 시작하기 || useNavidate( )',
@@ -16,11 +17,7 @@ export const metadata: Metadata = {
   },
 };
 
-type Props = {
-  searchParams: { [key: string]: string | undefined };
-};
-
-const Page = async ({ searchParams }: Props) => {
+const Page = async () => {
   const token = getServerCookie('access_token');
 
   if (!token) {
@@ -32,47 +29,13 @@ const Page = async ({ searchParams }: Props) => {
     return <LoginPrompt authStatus="unauthenticated" />;
   }
 
-  const backendUrl =
-    process.env.NEXT_PUBLIC_BACKEND_URL || 'http://localhost:3001';
-
-  const res = await fetch(`${backendUrl}/couples`, {
-    headers: { Authorization: `Bearer ${token}` },
-    cache: 'no-store',
-  });
-
-  if (!res.ok) {
-    throw new Error('커플 데이터 불러오기 실패');
-  }
-
-  const result = await res.json();
-  // console.log('커플 목록:', result);
-
-  if (!result.success || !Array.isArray(result.data)) {
-    throw new Error('잘못된 커플 데이터 형식');
-  }
-
-  const myCouple = result.data.find(
-    (c: any) => c.userAId === Number(userId) || c.userBId === Number(userId),
-  );
-
-  // console.log('내 커플:', myCouple);
+  const myCouple = await getMyCouple(token, Number(userId));
 
   if (myCouple) {
     redirect(`/date-calendar/${myCouple.id}`);
   }
 
-  return <LoginPrompt authStatus="no-couple" userId={userId} />;
+  return <LoginPrompt authStatus="no-couple" />;
 };
 
 export default Page;
-
-// // 테스트용 상수 (개발 환경에서만 사용)
-// const TEST_COUPLE_ID = '1';
-// const TEST_START_DATE = '2024-01-01';
-// const TEST_USER_ID = '1';
-
-// const page = () => {
-//   return <CalendarTabs coupleId={TEST_COUPLE_ID} startDate={TEST_START_DATE} userId={TEST_USER_ID} />;
-// };
-
-// export default page;
